WR 12262: Port Soar to work on Solaris.
This is a fix for another static initialization problem for wxIFM on wxGTK. The previous commit (12 May) was published on the wxWidgets Discussion Forum, but this one I found and fixed myself. Without this fix, Soar will run, but it spews out a bunch of messages labeled either Gtk-CRITICAL or GLib-GOjbect-CRITICAL with failed assertions.

git-svn-id: https://isrsvn.usurf.usu.edu/svn/isr/trunk/Contrib/wxIFM@2634 a06e8daa-702c-0410-a579-eb5e6c4fa09d